build.bat building a distribution kit

+++ plugin/compile_all_jsmblh.py
@@ -1,4 +1,4 @@
-import os, sys, re, glob
+import os, sys, re, glob, traceback
 import settings, store
 
 
@@ -11,15 +11,20 @@ def recompile(pluginpath = os.path.abspath(os.path.split(sys.argv[0])[0]), verbo
     for fn in glob.glob(os.path.join(pluginpath, '..', 'plugin', '*.py')):
         if ('def' + ' get_balance(') in open(fn, encoding='utf8').read():
             plugin = os.path.splitext(os.path.split(fn)[1])[0]
-            fl = 'p_' + plugin
-            module = __import__(plugin, globals(), locals(), [], 0)
-            data = tmpl.replace('{{pluginname}}', fl).replace('{{port}}', port)
-            if hasattr(module,'icon'):
-                data = re.sub(r'//\s*Icon\s*:\s*\S*', f'// Icon      : {module.icon}', data)
-            plugin_name = os.path.join(pluginpath, '..','jsmblhplugin', fl+'_localweb.jsmb')
-            open(plugin_name, 'w').write(data)
-            if verbose:
-                print(os.path.abspath(plugin_name))
+            try:
+                fl = 'p_' + plugin
+                module = __import__(plugin, globals(), locals(), [], 0)
+                data = tmpl.replace('{{pluginname}}', fl).replace('{{port}}', port)
+                if hasattr(module,'icon'):
+                    data = re.sub(r'//\s*Icon\s*:\s*\S*', f'// Icon      : {module.icon}', data)
+                plugin_name = os.path.join(pluginpath, '..','jsmblhplugin', fl+'_localweb.jsmb')
+                open(plugin_name, 'w').write(data)
+                if verbose:
+                    print(os.path.abspath(plugin_name))
+            except Exception:
+                errmsg = "".join(traceback.format_exception(*sys.exc_info()))
+                if 'import tkinter' not in errmsg:
+                    print(f'Compile {plugin} fail: {errmsg}')
 
 
 if __name__ == '__main__':
